+++ lib/index.ts
@@ -1,0 +1,3 @@
+import PersistentStorage = require('./PersistentStorage')
+export = PersistentStorage
+