+++ typings/tsd.d.ts
@@ -1,0 +1,5 @@
+/// <reference path="lodash/lodash.d.ts" />
+/// <reference path="node/node.d.ts" />
+/// <reference path="lz-string/lz-string.d.ts" />
+/// <reference path="chance/chance.d.ts" />
+/// <reference path="mocha/mocha.d.ts" />